GenBC coupling in FSI and effective direction (#352)
* Area updated each time for coupled face between 3D and 0D

Recalculate the area coupled between 0D and 3D at each time step. When effective direction is applied for this surface, the area will change over time step and the pressure should be calculated based on the new area.

* Improve file path for GenBC coupling

Improve the file path finding in genBC. Original case failed to find the communication file with FSI equation.